+++ portal-ui/src/common/SecureComponent/permissions.ts
@@ -163,6 +163,7 @@ export const IAM_PAGES = {
   TIERS_ADD_SERVICE: "/settings/tiers/add/:service",
   SITE_REPLICATION: "/settings/site-replication",
   SITE_REPLICATION_STATUS: "/settings/site-replication/status",
+  SITE_REPLICATION_ADD: "/settings/site-replication/add",
 
   /* Operator */
   TENANTS: "/tenants",
@@ -386,6 +387,10 @@ export const IAM_PAGES_PERMISSIONS = {
     IAM_SCOPES.ADMIN_CONFIG_UPDATE,
   ],
   [IAM_PAGES.SITE_REPLICATION_STATUS]: [
+    IAM_SCOPES.ADMIN_SERVER_INFO,
+    IAM_SCOPES.ADMIN_CONFIG_UPDATE,
+  ],
+  [IAM_PAGES.SITE_REPLICATION_ADD]: [
     IAM_SCOPES.ADMIN_SERVER_INFO,
     IAM_SCOPES.ADMIN_CONFIG_UPDATE,
   ],
